requirements.txt
matplotlib
opencv-python
pillow
numpy
jupyter
paddleocr==2.7.3
paddlepaddle
streamlit
wandb
python-dotenv